Fix 3D rendering (Atleast in my modified version)
I noticed you moved the gl.Clear and gl.ClearColor functions to a META:Clear function in framebuffer.lua, I didn't bother to check where this is called I just added the lines into the META:Begin function because I know it fixes it. Hopefully this is all that fixed it in mine, if its not let me know. Mine is heavily modified.